(WIP) Login + setting screen are now working
