+++ package.json
@@ -22,11 +22,13 @@
     "v-clipboard": "^2.2.2",
     "vue": "^2.6.10",
     "vue-js-modal": "^1.3.31",
+    "vue-router": "^3.3.4",
     "vue-toasted": "^1.1.27",
     "webextension-polyfill": "^0.3.1"
   },
   "devDependencies": {
     "@babel/core": "^7.1.2",
+    "@babel/plugin-proposal-class-properties": "^7.10.4",
     "@babel/plugin-proposal-optional-chaining": "^7.0.0",
     "@babel/preset-env": "^7.1.0",
     "@babel/runtime-corejs3": "^7.4.0",
