+++ webpack.config.js
@@ -30,6 +30,12 @@ const config = {
         test: /\.js$/,
         loader: 'babel-loader',
         exclude: /node_modules/,
+        options: {
+            presets: [
+              {'plugins': ['@babel/plugin-proposal-class-properties']
+            }
+          ]
+        }
       },
       {
         test: /\.css$/,
